+++ src/creditNote/invoice/discount.php
@@ -1,8 +1,8 @@
 <?php
 
 namespace StarkInfra\CreditNote\Invoice;
-use StarkInfra\Utils\Checks;
-use StarkInfra\Utils\SubResource;
+use StarkCore\Utils\Checks;
+use StarkCore\Utils\SubResource;
 
 
 class Discount extends SubResource